gan choi duoc roi

+++ src/uet/oop/bomberman/entities/StillEntity/Grass.java
@@ -9,7 +9,7 @@ import uet.oop.bomberman.entities.Entity;
 
 public class Grass extends StillEntity {
 
-    public Grass(int x, int y, Image img) {
+    public Grass(double x, double y, Image img) {
         super(x, y, img);
     }
